RA-1290 Using double for lats and longs instead of decimal.

+++ src/Esfa.Vacancy.Api.Types/GeoPoint.cs
@@ -5,11 +5,11 @@
         /// <summary>
         /// The longitude.
         /// </summary>
-        public decimal? Longitude { get; set; }
+        public double? Longitude { get; set; }
 
         /// <summary>
         /// The latitude.
         /// </summary>
-        public decimal? Latitude { get; set; }
+        public double? Latitude { get; set; }
     }
 }
